Project Reorganization
This build will support multiple Kalix and Akka services.
The inventory service was placed in "inventory" and the
organization service similarly. Customer is "practice" from
learning Kalix and will be converted to the "Member" service
eventually.

A project/Helpers.scala file was created to provide build helper
functions, most notably the "kalix" configuration function which
is used for the organization subproject in build.sbt